WIP: Concept of allowing bytes from series to be "stolen"
This makes any dynamic binary series add a little extra space (size of
a pointer) so that it may be "stolen" from Rebol.  Then it may also
be given back.  This parallels the trick used by rebMalloc, by making
every binary series pay for this when allocated.

The concept got a little ways in implementation but needs to be thought
about some more.