GHOST! vars unset, solidify Level_At_End()
There was a concept that got introduced by skipping commas that was not
a good idea, which was that you couldn't test a-prior for if a level
was at its end.  This was tied up with some other bad ideas about the
notion of a "true unset" in the dual band.  All of that had kind of
been debunked in the past, but there was an idea that dual states might
change the reality--but they do not.

This goes with a much better reality: the use of GHOST! as the "unset"
state, distinct from VOID and TRASH!.  This is more viable now that
GHOST! handling has become safe and easy.