^void workaround to bridge "unsetness" ideas
GHOST! has settled on being the antiform that is held by a variable
that is considered to be "truly unset" (well, as unset as a variable
can be).  If you do a non-^META assignment of either a ghost or a trash
then it will turn a variable into a ghost, now.

    x: ()  ; stores ghost, as if you'd written ^x: ()
    x: opt null  ; also stores ghost, despite writing void

This gives two convenient directions for "unsetting" things.  But
neither of those assign trash, which is the bootstrap executable's
notion of "unset", and it's mechanically prohibited to set variables
to its idea of void.

This bridges the gap with:

   x: ^void

It looks leftward, and if there's a set-word there it will set it
to trash.  if not, it will just evaluate to void.

This allows ^void to unset a variable via assignment in both the
new executable and the bootstrap executable--an important idea for
when a variable is holding a function.